requirements.txt
colorama
numpy
python-dotenv